# Perf: Speed-up `humanize`