@Ptico 2017-11-25
Very clever, thanks!